# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Maz Delivery API",
        "description": "Maz Delivery project",
        "contact": {
            "email": "info@mazdelivery.com"
        },
        "license": {
            "name": "Proprietary",
            "url": "http://www.mazdelivery.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/bank/{id}": {
            "get": {
                "tags": [
                    "bank"
                ],
                "summary": "Show an Bank",
                "description": "Show an Bank",
                "operationId": "bankDetails",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Bank ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "bank"
                ],
                "summary": "Update an Bank",
                "description": "Update an Bank",
                "operationId": "bankUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Bank ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BankUpdateFormRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "bank"
                ],
                "summary": "Delete an Bank",
                "description": "Delete an Bank",
                "operationId": "bankDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Bank ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/bank-details": {
            "get": {
                "tags": [
                    "bankDetails"
                ],
                "summary": "Get All Bank Details",
                "description": "Get All Bank Details",
                "operationId": "bankdetailsLists",
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "bankDetails"
                ],
                "summary": "Post New Bank Details",
                "description": "Post New Bank Details",
                "operationId": "bankDetailsCreate",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BankDetailCreateFormRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/bank-details/{id}": {
            "get": {
                "tags": [
                    "bankDetails"
                ],
                "summary": "Show an Bank Details",
                "description": "Show an Bank Details",
                "operationId": "bankDetailsDetails",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Bank Detail ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "bankDetails"
                ],
                "summary": "Update an Bank Details",
                "description": "Update an Bank Details",
                "operationId": "bankDetailsUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Bank Detail ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BankDetailUpdateFormRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "bankDetails"
                ],
                "summary": "Delete an Bank Details",
                "description": "Delete an Bank Details",
                "operationId": "bankDetailsDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Bank Detail ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "BankDetailCreateFormRequest": {
                "title": "Bank Detail Create Form Request Fields",
                "description": "Bank Detail Create request body data",
                "required": [
                    "first_name"
                ],
                "type": "object"
            },
            "BankDetailUpdateFormRequest": {
                "title": "Bank Detail Update Form Request Fields",
                "description": "Bank Detail Update request body data",
                "required": [
                    "first_name"
                ],
                "type": "object"
            },
            "BankUpdateFormRequest": {
                "title": "Bank Update Form Request Fields",
                "description": "Bank Update request body data",
                "required": [
                    "first_name"
                ],
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}